rename _eod to _historical in standard_models
