+++ openbb_sdk/providers/fmp/openbb_fmp/models/forex_eod.py
@@ -6,7 +6,7 @@ from typing import Any, Dict, List, Optional
 
 from dateutil.relativedelta import relativedelta
 from openbb_provider.abstract.fetcher import Fetcher
-from openbb_provider.standard_models.forex_eod import (
+from openbb_provider.standard_models.forex_historical import (
     ForexHistoricalData,
     ForexHistoricalQueryParams,
 )
